+++ aim/web/ui/src/components/ImagesSet/ImagesSet.d.ts
@@ -2,6 +2,7 @@ import { IFocusedState } from 'types/services/models/metrics/metricsAppModel';
 
 export interface IImageSetProps {
   data: any;
+  orderedMap: { [key: string]: any };
   imagesBlobs: object;
   onScroll: () => void;
   addUriToList: (blobUrl: string) => void;